% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{col_labels}
\alias{col_labels}
\title{generate column labels for a micro titre plate}
\usage{
col_labels(cols = NULL, n = NULL)
}
\arguments{
\item{cols}{<num> vector of column indices}

\item{n}{<int> number of columns in micro titre plate}
}
\description{
Specify the col indices for which to get labels or the number of cols
}
\examples{
    col_labels(cols = 3)
    col_labels(n = 3L) # same as col_labels(cols = 1:3)

}
