% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guess_dims_from_n}
\alias{guess_dims_from_n}
\title{guess the dimensions of microtitre plate from the number of wells.}
\usage{
guess_dims_from_n(n)
}
\arguments{
\item{n}{`<num>` number of wells}
}
\value{
[list()] of length 2: number of rows (`nrow`) and columns (`ncol`).
}
\description{
guess the dimensions of microtitre plate from the number of wells.
}
\keyword{internal}
