% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{row_labels}
\alias{row_labels}
\title{generate row labels for a micro titre plate}
\usage{
row_labels(rows = NULL, n = NULL)
}
\arguments{
\item{rows}{<num> vector of row indices}

\item{n}{<int> number of rows in micro titre plate}
}
\description{
Specify the row indices for which to get labels or the number of rows.
}
\examples{
    row_labels(rows = 3)
    row_labels(n = 3L) # same as row_labels(rows = 1:3)

}
