% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{well_labels}
\alias{well_labels}
\title{well labels from dimensions of microtitre plate.}
\usage{
well_labels(nrow, ncol, byrow = TRUE, rownames = row_labels(n = nrow),
  colnames = col_labels(n = ncol), nchar_row = max(nchar(rownames)),
  nchar_col = max(nchar(colnames)))
}
\arguments{
\item{nrow}{`<int>` number of rows}

\item{ncol}{`<int>` number of cols}

\item{byrow}{`<lgl>` defaults to `"rowmajor"` [TRUE], `"colmajor"` use [FALSE]}

\item{rownames}{`<chr>` row names to use, defaults to `row_labels(n = nrow)`}

\item{colnames}{`<chr>` col names to use, defaults to `col_labels(n = ncol)`}

\item{col_nchar}{`<int>` the number of characters for the column section. This
controls the left-padding of well labels with zeros. e.g. A1 vs. A01. The
default is the number of characters in `ncol` for nice lexicographic
sorting behaviour.}
}
\value{
`<chr>` vector of well labels where the row is represented by a letter,
 and the col is represented by an intenger, possibly left-padded with one
 or more zeros.
}
\description{
well labels from dimensions of microtitre plate.
}
