% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{well_labels_from_length}
\alias{well_labels_from_length}
\title{get the labels for a microtitre plate based on the length of the vector.}
\usage{
well_labels_from_length(wells = NULL, n = NULL)
}
\arguments{
\item{wells}{vector of wells}

\item{n}{number of wells}
}
\description{
get the labels for a microtitre plate based on the length of the vector.
}
